requirements.txt
aiohttp
asyncio
ph4_walkingpad
psycopg2-binary
pyyaml
